Correcciones paquete

+++ src/database/migrations/0000_00_00_000001_nodes_product.php
@@ -12,19 +12,21 @@ class NodesProduct extends Migration
      */
     public function up()
     {
-        Schema::create('variations', function (Blueprint $table) {
-            $table->increments('id');
-            $table->enum('type', ['normal', 'color', 'image'])->nullable()->default('normal');
-            $table->timestamps();
-        });
-        Schema::create('variation_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('variation_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['variation_id','locale']);
-            $table->foreign('variation_id')->references('id')->on('variations')->onDelete('cascade');
-        });
+        if(config('product.product_variations')){
+            Schema::create('variations', function (Blueprint $table) {
+                $table->increments('id');
+                $table->enum('type', ['normal', 'color', 'image'])->nullable()->default('normal');
+                $table->timestamps();
+            });
+            Schema::create('variation_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('variation_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['variation_id','locale']);
+                $table->foreign('variation_id')->references('id')->on('variations')->onDelete('cascade');
+            });
+        }
         Schema::create('categories', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('parent_id')->nullable();
@@ -50,24 +52,32 @@ class NodesProduct extends Migration
             $table->string('image')->nullable();
             $table->boolean('printed')->nullable()->default(0);
             $table->integer('currency_id')->unsigned();
-            $table->integer('partner_id')->nullable();
-            $table->integer('partner_transport_id')->nullable();
-            $table->integer('external_currency_id')->unsigned();
-            $table->decimal('currency_product_cost', 10, 2)->nullable();
-            $table->decimal('currency_transport_cost', 10, 2)->nullable();
             $table->decimal('weight', 10, 2)->nullable()->default(0);
-            $table->decimal('exchange', 10, 2)->nullable();
-            $table->decimal('product_cost', 10, 2)->nullable();
-            $table->decimal('transport_cost', 10, 2)->nullable();
             $table->decimal('cost', 10, 2)->nullable();
             $table->decimal('price', 10, 2)->nullable();
-            $table->decimal('no_invoice_price', 10, 2)->nullable();
-            $table->decimal('offer_price', 10, 2)->nullable();
-            $table->integer('product_group_id')->nullable();
             $table->timestamps();
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->foreign('currency_id')->references('id')->on('currencies')->onDelete('cascade');
         });
+        if(config('product.product_extras')){
+            Schema::table('products', function (Blueprint $table) {
+                $table->integer('partner_id')->nullable();
+                $table->integer('partner_transport_id')->nullable();
+                $table->integer('external_currency_id')->unsigned();
+                $table->decimal('currency_product_cost', 10, 2)->nullable();
+                $table->decimal('currency_transport_cost', 10, 2)->nullable();
+                $table->decimal('exchange', 10, 2)->nullable();
+                $table->decimal('product_cost', 10, 2)->nullable();
+                $table->decimal('transport_cost', 10, 2)->nullable();
+                $table->decimal('no_invoice_price', 10, 2)->nullable();
+                $table->decimal('offer_price', 10, 2)->nullable();
+            });
+        }
+        if(config('product.product_groups')){
+            Schema::table('products', function (Blueprint $table) {
+                $table->integer('product_group_id')->nullable();
+            });
+        }
         Schema::create('product_translation', function(Blueprint $table) {
             $table->increments('id');
             $table->integer('product_id')->unsigned();
@@ -76,92 +86,104 @@ class NodesProduct extends Migration
             $table->unique(['product_id','locale']);
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
         });
-        Schema::create('product_groups', function (Blueprint $table) {
-            $table->increments('id');
-            $table->timestamps();
-        });
-        Schema::create('product_group_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('product_group_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['product_group_id','locale']);
-            $table->foreign('product_group_id')->references('id')->on('product_groups')->onDelete('cascade');
-        });
-        Schema::create('product_variation', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('product_id')->unsigned();
-            $table->integer('variation_id')->unsigned();
-            $table->string('value')->nullable();
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->foreign('variation_id')->references('id')->on('variations')->onDelete('cascade');
-        });
-        Schema::create('product_benefits', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned();
-            $table->string('name')->nullable();
-            $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
-        });
-        Schema::create('product_benefit_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('product_benefit_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['product_benefit_id','locale']);
-            $table->foreign('product_benefit_id')->references('id')->on('product_benefits')->onDelete('cascade');
-        });
-        Schema::create('product_images', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned();
-            $table->string('image')->nullable();
-            $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
-        });
-        Schema::create('product_image_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('product_image_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['product_image_id','locale']);
-            $table->foreign('product_image_id')->references('id')->on('product_images')->onDelete('cascade');
-        });
-        Schema::create('product_offers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned();
-            $table->enum('type', ['discount_percentage','discount_value','discount_quantities'])->nullable()->default('discount_percentage');
-            $table->string('value')->nullable();
-            $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
-        });
-        Schema::create('product_offer_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('product_offer_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['product_offer_id','locale']);
-            $table->foreign('product_offer_id')->references('id')->on('product_offers')->onDelete('cascade');
-        });
-        Schema::create('packages', function (Blueprint $table) {
-            $table->increments('id');
-            $table->boolean('active')->nullable()->default(1);
-            $table->integer('currency_id')->unsigned();
-            $table->string('price')->nullable();
-            $table->foreign('currency_id')->references('id')->on('currencies')->onDelete('cascade');
-        });
-        Schema::create('package_translation', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('package_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('name')->nullable();
-            $table->unique(['package_id','locale']);
-            $table->foreign('package_id')->references('id')->on('packages')->onDelete('cascade');
-        });
-        Schema::create('package_products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned();
-            $table->integer('product_id')->unsigned();
-            $table->string('quantity')->nullable();
-            $table->foreign('parent_id')->references('id')->on('packages')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-        });
+        if(config('product.product_groups')){
+            Schema::create('product_groups', function (Blueprint $table) {
+                $table->increments('id');
+                $table->timestamps();
+            });
+            Schema::create('product_group_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('product_group_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['product_group_id','locale']);
+                $table->foreign('product_group_id')->references('id')->on('product_groups')->onDelete('cascade');
+            });
+        }
+        if(config('product.product_variations')){
+            Schema::create('product_variation', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('product_id')->unsigned();
+                $table->integer('variation_id')->unsigned();
+                $table->string('value')->nullable();
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+                $table->foreign('variation_id')->references('id')->on('variations')->onDelete('cascade');
+            });
+        }
+        if(config('product.product_benefits')){
+            Schema::create('product_benefits', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('parent_id')->unsigned();
+                $table->string('name')->nullable();
+                $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
+            });
+            Schema::create('product_benefit_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('product_benefit_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['product_benefit_id','locale']);
+                $table->foreign('product_benefit_id')->references('id')->on('product_benefits')->onDelete('cascade');
+            });
+        }
+        if(config('product.product_images')){
+            Schema::create('product_images', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('parent_id')->unsigned();
+                $table->string('image')->nullable();
+                $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
+            });
+            Schema::create('product_image_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('product_image_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['product_image_id','locale']);
+                $table->foreign('product_image_id')->references('id')->on('product_images')->onDelete('cascade');
+            });
+        }
+        if(config('product.product_offers')){
+            Schema::create('product_offers', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('parent_id')->unsigned();
+                $table->enum('type', ['discount_percentage','discount_value','discount_quantities'])->nullable()->default('discount_percentage');
+                $table->string('value')->nullable();
+                $table->foreign('parent_id')->references('id')->on('products')->onDelete('cascade');
+            });
+            Schema::create('product_offer_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('product_offer_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['product_offer_id','locale']);
+                $table->foreign('product_offer_id')->references('id')->on('product_offers')->onDelete('cascade');
+            });
+        }
+        if(config('product.product_packages')){
+            Schema::create('packages', function (Blueprint $table) {
+                $table->increments('id');
+                $table->boolean('active')->nullable()->default(1);
+                $table->integer('currency_id')->unsigned();
+                $table->string('price')->nullable();
+                $table->foreign('currency_id')->references('id')->on('currencies')->onDelete('cascade');
+            });
+            Schema::create('package_translation', function(Blueprint $table) {
+                $table->increments('id');
+                $table->integer('package_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('name')->nullable();
+                $table->unique(['package_id','locale']);
+                $table->foreign('package_id')->references('id')->on('packages')->onDelete('cascade');
+            });
+            Schema::create('package_products', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('parent_id')->unsigned();
+                $table->integer('product_id')->unsigned();
+                $table->string('quantity')->nullable();
+                $table->foreign('parent_id')->references('id')->on('packages')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
     }
 
     /**
