+++ src/database/seeds/TruncateSeeder.php
@@ -13,22 +13,34 @@ class TruncateSeeder extends Seeder {
      */
     public function run()
     {
-        \Solunes\Product\App\PackageProduct::truncate();
-        \Solunes\Product\App\PackageTranslation::truncate();
-        \Solunes\Product\App\Package::truncate();
-        \Solunes\Product\App\ProductOfferTranslation::truncate();
-        \Solunes\Product\App\ProductOffer::truncate();
-        \Solunes\Product\App\ProductImageTranslation::truncate();        
-        \Solunes\Product\App\ProductImage::truncate();
-        \Solunes\Product\App\ProductBenefitTranslation::truncate();
-        \Solunes\Product\App\ProductBenefit::truncate();
-        \Solunes\Product\App\ProductGroupTranslation::truncate();
-        \Solunes\Product\App\ProductGroup::truncate();
+        if(config('product.product_packages')){
+            \Solunes\Product\App\PackageProduct::truncate();
+            \Solunes\Product\App\PackageTranslation::truncate();
+            \Solunes\Product\App\Package::truncate();
+        }
+        if(config('product.product_offers')){
+            \Solunes\Product\App\ProductOfferTranslation::truncate();
+            \Solunes\Product\App\ProductOffer::truncate();
+        }
+        if(config('product.product_images')){
+            \Solunes\Product\App\ProductImageTranslation::truncate();        
+            \Solunes\Product\App\ProductImage::truncate();
+        }
+        if(config('product.product_benefits')){
+            \Solunes\Product\App\ProductBenefitTranslation::truncate();
+            \Solunes\Product\App\ProductBenefit::truncate();
+        }
+        if(config('product.product_groups')){
+            \Solunes\Product\App\ProductGroupTranslation::truncate();
+            \Solunes\Product\App\ProductGroup::truncate();
+        }
         \Solunes\Product\App\ProductTranslation::truncate();
         \Solunes\Product\App\Product::truncate();
         \Solunes\Product\App\CategoryTranslation::truncate();
         \Solunes\Product\App\Category::truncate();
-        \Solunes\Product\App\VariationTranslation::truncate();
-        \Solunes\Product\App\Variation::truncate();
+        if(config('product.product_variations')){
+            \Solunes\Product\App\VariationTranslation::truncate();
+            \Solunes\Product\App\Variation::truncate();
+        }
     }
 }
